Change API url for /api/compliance/arf_reports
Fixes #89.
