+++ app/controllers/api/v2/compliance/arf_reports_controller.rb
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2014 Red Hat Inc.
+# Copyright (c) 2014--2015 Red Hat Inc.
 #
 # This software is licensed to you under the GNU General Public License,
 # version 3 (GPLv3). There is NO WARRANTY for this software, express or
@@ -21,7 +21,43 @@ module Api
 
         add_puppetmaster_filters :create
 
-        api :POST, "/arf/:cname/:policy_id/:date", N_("Upload an ARF report")
+        before_filter :find_resource, :only => %w{show destroy}
+
+        def resource_name
+          'Scaptimony::ArfReport'
+        end
+
+        def get_resource
+          instance_variable_get :"@arf_report" or raise 'no resource loaded'
+        end
+
+        resource_description do
+          resource_id 'scaptimony_arf_reports'
+          api_version 'v2'
+          api_base_url "/api/v2"
+        end
+
+        api :GET, '/compliance/arf_reports', N_('List Arf reports')
+        param_group :search_and_pagination, ::Api::V2::BaseController
+
+        def index
+          @arf_reports = resource_scope_for_index(:permission => :edit_compliance).includes(:arf_report_breakdown, :asset)
+        end
+
+        api :GET, '/compliance/arf_reports/:id', N_('Show an Arf report')
+        param :id, :identifier, :required => true
+
+        def show
+        end
+
+        api :DELETE, '/compliance/arf_reports/:id', N_('Deletes an Arf Report')
+        param :id, :identifier, :required => true
+
+        def destroy
+          process_response @arf_report.destroy
+        end
+
+        api :POST, "/compliance/arf/:cname/:policy_id/:date", N_("Upload an ARF report")
         param :cname, :identifier, :required => true
         param :policy_id, :identifier, :required => true
         param :date, :identifier, :required => true
@@ -43,8 +79,15 @@ module Api
           # no matter what content-encoding says. Let's pass content-type arf-bzip2
           # and move forward.
           super unless
+            params[:action] == 'create' and
             request.content_type.end_with? 'arf-bzip2' and
             request.env['HTTP_CONTENT_ENCODING'] == 'x-bzip2'
+        end
+
+        private
+        def find_resource
+          not_found and return if params[:id].blank?
+          instance_variable_set("@arf_report", resource_scope.find(params[:id]))
         end
       end
     end
