--- app/views/api/v2/arf_reports/index.json.rabl
@@ -1,2 +1,0 @@
-collection @arf_reports
-extends "api/v2/arf_reports/main"
