--- app/views/api/v2/arf_reports/main.json.rabl
@@ -1,5 +1,0 @@
-object @arf_report
-
-extends "api/v2/arf_reports/base"
-
-attributes :created_at, :updated_at
