--- app/views/api/v2/arf_reports/show.json.rabl
@@ -1,3 +1,0 @@
-object @arf_report
-
-extends "api/v2/arf_reports/main"
