+++ app/views/api/v2/compliance/arf_reports/base.json.rabl
