+++ app/views/api/v2/compliance/arf_reports/index.json.rabl
@@ -1,0 +1,2 @@
+collection @arf_reports
+extends "api/v2/compliance/arf_reports/main"
