+++ app/views/api/v2/compliance/arf_reports/main.json.rabl
@@ -1,0 +1,5 @@
+object @arf_report
+
+extends "api/v2/compliance/arf_reports/base"
+
+attributes :created_at, :updated_at
