+++ app/views/api/v2/compliance/arf_reports/show.json.rabl
@@ -1,0 +1,3 @@
+object @arf_report
+
+extends "api/v2/compliance/arf_reports/main"
